# src/main/java/com/cozi/soft/listing/core/model/entity/ListingAddress.java
package com.cozi.soft.listing.core.model.entity;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.Table;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.experimental.Accessors;
import lombok.experimental.SuperBuilder;

@Getter
@Setter
@Entity
@NoArgsConstructor
@AllArgsConstructor
@Accessors(chain = true)
@SuperBuilder(toBuilder = true)
@Table(name = "listing_address")
public class ListingAddress extends BaseEntity {
    @Column(name = "country", nullable = false)
    private String country;
    @Column(name = "state", nullable = false)
    private String state;
    @Column(name = "city", nullable = false)
    private String city;
    private String zipCode;
    private String street;

//    @Column(name = "location", columnDefinition = "GEOMETRY('POINT', 4326)")
//    private Geometry location;
}